@@ -5,13 +5,13 @@ repository: https://github.com/blaugold/flutter_exploded
 issue_tracker: https://github.com/blaugold/flutter_exploded/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  vm_service: ^4.1.0
+  vm_service: ^7.5.0
 
 dev_dependencies:
   flutter_test:
